core: the doors a reader writes structure through
The facade is read-only, and the import lint says the same a second way: outside
the core only the facade may be imported. So a reader that rebuilds a graph from
a file writes through the graph object, which is inside the core and may reach
the build and mutation layers freely.

Six doors carry it. Installing a whole graph, setting the kind of each entity a
map names, moving entities to new keys, setting one field of one edge, turning an
edge into a hyperedge, and setting the coefficients of a column outright. The
install also takes no matrix, for a file that stores none, and defers the
adjacency indexes, because most loads never ask an adjacency question.

Three defects came out of routing the readers, and each was invisible while the
two stores were compared only against each other.

A coefficient lost precision. The canonical arrays held float32, which is the
precision of the matrix and not of a record, so one third came back as
0.3333333432674408. Both arrays are double now and the matrix layer casts where
it materializes.

A hyperedge with no target side was reported as directed. The record store
answers that from whether a target side exists at all, which is what makes a
boundary reaction an undirected member set. The slot store now reads the member
roles, which already say so.

An endpoint the graph met for the first time inside a bulk edge add never reached
the slot store, so a graph built from edges alone had no nodes.

Also: the coefficients of an edge are answered in the form the caller holds its
own ids in, as the sides of an edge already were.

+++ annnet/core/_store.py
@@ -120,7 +120,7 @@ class CoreState:
         # Edge arrays, indexed by edge slot.
         self.edge_kind = np.zeros(_INITIAL_CAPACITY, dtype=np.uint8)
         self.edge_directed = np.full(_INITIAL_CAPACITY, INHERIT, dtype=np.int8)
-        self.edge_weight = np.ones(_INITIAL_CAPACITY, dtype=np.float32)
+        self.edge_weight = np.ones(_INITIAL_CAPACITY, dtype=np.float64)
         self.edge_explicit = np.zeros(_INITIAL_CAPACITY, dtype=bool)
 
         # Member lists. Each edge owns a segment of the three pools. A delete
@@ -128,7 +128,7 @@ class CoreState:
         self.member_start = np.zeros(_INITIAL_CAPACITY, dtype=np.int64)
         self.member_len = np.zeros(_INITIAL_CAPACITY, dtype=np.int32)
         self.member_ent = np.zeros(_INITIAL_CAPACITY, dtype=np.int64)
-        self.member_coef = np.zeros(_INITIAL_CAPACITY, dtype=np.float32)
+        self.member_coef = np.zeros(_INITIAL_CAPACITY, dtype=np.float64)
         self.member_role = np.zeros(_INITIAL_CAPACITY, dtype=np.int8)
         self._member_used = 0
 
@@ -531,7 +531,7 @@ class CoreState:
         before = self._member_used
         total = int(self.member_len[slots].sum()) if slots.size else 0
         ent = np.zeros(max(_INITIAL_CAPACITY, total), dtype=np.int64)
-        coef = np.zeros(max(_INITIAL_CAPACITY, total), dtype=np.float32)
+        coef = np.zeros(max(_INITIAL_CAPACITY, total), dtype=np.float64)
         role = np.zeros(max(_INITIAL_CAPACITY, total), dtype=np.int8)
         cursor = 0
         for slot in slots:
